Array fix and ShortPointer added

+++ src/test/resources/io/github/nowipi/ffm/array.c
@@ -2,12 +2,12 @@
 
 typedef struct {
     int *data;
-    long count;
+    size_t count;
 } array_t;
 
-__declspec(dllexport) array_t *array_new(long elementCount);
+__declspec(dllexport) array_t *array_new(size_t elementCount);
 
-array_t *array_new(long elementCount) {
+array_t *array_new(size_t elementCount) {
     array_t *array = (array_t *) malloc(sizeof(array_t));
     array->data = (int *) calloc(elementCount, sizeof(int));
     array->count = elementCount;
